requirements.txt
grpcio==1.60.1
joblib==1.3.2
numpy==1.26.4
pandas==2.0.0
protobuf==4.25.2
python-dotenv==1.0.1
scikit_learn==1.4.0
seaborn==0.13.2
pymongo==4.6.2
redis==5.0.3